requirements.txt
appdirs==1.4.3
appnope==0.1.0
decorator==4.0.11
dj-database-url==0.4.2
Django==1.11.1
django-simple-email-confirmation==0.22
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==3.3.0
